# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/recipes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets a list of recipes based on the provided filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Gets a list of recipes based on the provided filters.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_recipes.Recipe"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_recipes.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_recipes.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Generates a new recipe using OpenAI if it is a new set of parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Generates a new recipe using OpenAI if it is a new set of parameters.",
                "parameters": [
                    {
                        "description": "Generate recipe request.",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_recipes.RecipeDefinitions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_recipes.Recipe"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_recipes.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_recipes.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github.com_rubengomes8_HappyMouthBackend_internal_recipes.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github.com_rubengomes8_HappyMouthBackend_internal_recipes.Recipe": {
            "type": "object",
            "properties": {
                "calories": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "definitions": {
                    "$ref": "#/definitions/github.com_rubengomes8_HappyMouthBackend_internal_recipes.RecipeDefinitions"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github.com_rubengomes8_HappyMouthBackend_internal_recipes.RecipeDefinitions": {
            "type": "object",
            "properties": {
                "exclude_ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "include_ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_recipes.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_recipes.Recipe": {
            "type": "object",
            "properties": {
                "calories": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "definitions": {
                    "$ref": "#/definitions/internal_recipes.RecipeDefinitions"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_recipes.RecipeDefinitions": {
            "type": "object",
            "properties": {
                "exclude_ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "include_ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}